Updated job to run on datasource creation
